// christmas-task/src/types/interfaces.ts
export interface IRangeFilterData {
  minValue: number;
  maxValue: number;
  step: number;
  title: string;
}

export interface IValueFilterData {
  typeValues: string[];
  buttonStyle: string;
  title: string;
}

export interface IToyData {
  num: string;
  name: string;
  count: string;
  year: string;
  shape: string;
  color: string;
  size: string;
  favorite: string;
}